CxxCH counting script
This script loops through all the genomes that we have downloaded from
JGI and counts proteins that have CxxCH motifs.  For each genome the
script outputs the total number of proteins, the number of proteins
with at least one CxxCH motifs, the number of proteins with 8 or more
CxxCH motifs, and the number of CxxCH motifs in the protein with the
most motifs.  The script will also output a fasta file with all the
proteins containing at least one CxxCH motifs.